+++ docs/changelog.rst
@@ -47,5 +47,8 @@ v0.1
 * Enumerate truth tables: ``next_inplace``, ``next``
   `#10 <https://github.com/msoeken/kitty/pull/10>`_
 
+* Compute co-factors: ``cofactor0_inplace``, ``cofactor0``, ``cofactor1_inplace``, ``cofactor1``
+  `#11 <https://github.com/msoeken/kitty/pull/11>`_
+
 * NPN canonization: ``exact_npn_canonization``, ``create_from_npn_config``
   `#8 <https://github.com/msoeken/kitty/pull/8>`_